% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_normal_v7.R
\name{convert_to_ggmcmc}
\alias{convert_to_ggmcmc}
\title{Converts MCMC samples into \code{ggmcmc} format}
\usage{
convert_to_ggmcmc(
  x,
  pattern = c("mean", "sigma", "beta", "alpha", "pois_means", "T"),
  include_warmup = FALSE
)
}
\arguments{
\item{x}{(mcmc_hmm_*) MCMC HMM object}

\item{pattern}{(character) pattern(s) with model parameters to be included in the output}

\item{include_warmup}{(logical) include warmup samples. By default \code{FALSE}}
}
\value{
data.frame compatible with functions from the \code{ggmcmc} package
}
\description{
This helper function converts MCMC samples into \code{ggmcmc} format
}
\examples{
# TODO
}
