% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_documentation.R
\docType{data}
\name{example_hmm_mcmc_gamma_poisson}
\alias{example_hmm_mcmc_gamma_poisson}
\title{Example of a Simulated Gamma-Poisson Model}
\format{
hmm_mcmc_gamma_poisson object
}
\usage{
example_hmm_mcmc_gamma_poisson
}
\description{
Example of a Simulated Gamma-Poisson Model
}
\examples{
# Data
hist(example_hmm_mcmc_gamma_poisson$data, 
     breaks = 50, xlab = "", main = "")

# Priors
example_hmm_mcmc_gamma_poisson$priors

# Model
example_hmm_mcmc_gamma_poisson
summary(example_hmm_mcmc_gamma_poisson)
}
\keyword{datasets}
