% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_normal.R
\name{generate_random_T}
\alias{generate_random_T}
\title{Generate a Random Transition Matrix}
\usage{
generate_random_T(n = 3)
}
\arguments{
\item{n}{(integer) dimension of a transition matrix}
}
\value{
random \code{n x n} transition matrix
}
\description{
This helper function generates a transition matrix at random for testing purposes
}
\details{
Uniform random numbers \eqn{[0,1]} are used to fill the matrix. Rows are then
normalized.
}
\examples{
mat_T <- generate_random_T(3)
mat_T

rowSums(mat_T)
}
