% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_poisson.R
\name{hmm_simulate_gamma_poisson_data}
\alias{hmm_simulate_gamma_poisson_data}
\title{Simulate Data Based on a Gamma-Poisson Model for a Hidden Markov Model Simulation}
\usage{
hmm_simulate_gamma_poisson_data(L, mat_T, betas, alpha)
}
\arguments{
\item{L}{(integer) number of simulations}

\item{mat_T}{(matrix) a square matrix with the initial state}

\item{betas}{(numeric) \code{rate} parameter in \code{\link{rgamma}} for emission probabilities}

\item{alpha}{(numeric) \code{shape} parameter in \code{\link{rgamma}} for emission probabilities}
}
\value{
Returns a data vector "data", the "true" hidden states "states" used to generate the data vector
and prior probability of states "pi".
}
\description{
Simulate Data Based on a Gamma-Poisson Model for a Hidden Markov Model Simulation
}
\examples{
mat_T <- rbind(c(1-0.01, 0.01, 0),
               c(0.01, 1-0.02, 0.01),
               c(0, 0.01, 1-0.01))
L <- 2^7
betas <- c(0.1, 0.3, 0.5)
alpha <- 1

sim_data <- hmm_simulate_gamma_poisson_data(L = L,
                                            mat_T = mat_T,
                                            betas = betas,
                                            alpha = alpha)
hist(sim_data$data, breaks = 40,
     main = "Histogram of Simulated Gamma-Poisson Data", xlab = "")
sim_data
}
