% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_poisson.R
\name{hmm_simulate_poisgamma_data}
\alias{hmm_simulate_poisgamma_data}
\title{Simulate a Hidden Markov Model based on Poisson-gamma data}
\usage{
hmm_simulate_poisgamma_data(L, mat_T, betas, alpha)
}
\arguments{
\item{L}{(integer) number of simulations}

\item{mat_T}{(matrix) a square matrix with the initial state}

\item{betas}{(numeric) \code{rate} parameter in \code{\link{rgamma}} for emission probabilities}

\item{alpha}{(numeric) \code{shape} parameter in \code{\link{rgamma}} for emission probabilities}
}
\value{
returns a data vector "data", the "true" hidden states "states" used to generate the data vector
and prior probability of states "pi".
}
\description{
Simulate a Hidden Markov Model based on Poisson-gamma data
}
\examples{
mat_T <- rbind(c(1-0.01,0.01,0),
               c(0.01,1-0.02,0.01),
               c(0,0.01,1-0.01))
L <- 2^10
betas <- c(0.1, 0.3, 0.5)
alpha <- 1

sim_data <- hmm_simulate_poisgamma_data(L = L,
                                        mat_T = mat_T,
                                        betas = betas,
                                        alpha = alpha)
hist(sim_data$data, breaks = 100, main = "Histogram")
}
