% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_normal_v7.R
\name{kullback_leibler_cont_appr}
\alias{kullback_leibler_cont_appr}
\title{Calculate a continuous approximation of the Kullback-Leibler divergence}
\usage{
kullback_leibler_cont_appr(p, q)
}
\arguments{
\item{p}{(numeric) probabilities}

\item{q}{(numeric) probabilities}
}
\value{
Numeric vector
}
\description{
Calculate a continuous approximation of the Kullback-Leibler divergence
}
\details{
TODO: write formula
}
\examples{
n <- 100
dist1 <- rnorm(n)
dist2 <- rnorm(n, mean = 0, sd =  5)
dist3 <- rnorm(n, mean = 2, sd =  10)

kullback_leibler_cont_appr(density(dist1)$y, density(dist2)$y)
kullback_leibler_cont_appr(density(dist1)$y, density(dist3)$y)
}
