% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_poisson_v2.R
\name{kullback_leibler_disc}
\alias{kullback_leibler_disc}
\title{Calculate a discrete approximation of the Kullback-Leibler divergence}
\usage{
kullback_leibler_disc(p, q)
}
\arguments{
\item{p}{(numeric) probabilities}

\item{q}{(numeric) probabilities}
}
\value{
Numeric vector
}
\description{
Calculate a discrete approximation of the Kullback-Leibler divergence
}
\details{
TODO: write formula
}
\examples{
# TODO
}
