% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_documentation.R
\docType{package}
\name{oHMMed-package}
\alias{oHMMed-package}
\title{oHMMed: Hidden Markov Models with Ordered Hidden States and Emission Densities with Applications in Population Genetics}
\description{
The oHMMed package contains an implementation of Hidden Markov Models with ordered hidden states and emission densities with applications in population genetics.
}
\references{
Inference of Genomic Landscapes using Ordered Hidden Markov Models with Emission Densities (oHMMed)
Claus Vogl, Mariia Karapetiants, Burçin Yildirim, Hrönn Kjartansdottir, Carolin Kosiol, Juraj Bergman, Michal Majka, Lynette Caitlin Mikula
bioRxiv 2023.06.26.546495; doi: \doi{10.1101/2023.06.26.546495}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LynetteCaitlin/oHMMed}
  \item \url{https://lynettecaitlin.github.io/oHMMed/}
  \item Report bugs at \url{https://github.com/LynetteCaitlin/oHMMed/issues}
}

}
\author{
\strong{Maintainer}: Michal Majka \email{michalmajka@hotmail.com} (\href{https://orcid.org/0000-0002-7524-8014}{ORCID})

Authors:
\itemize{
  \item Lynette Caitlin Mikula \email{lynettecaitlin@gmail.com} (\href{https://orcid.org/0000-0002-0252-4014}{ORCID})
  \item Claus Vogl \email{claus.vogl@vetmeduni.ac.at} (\href{https://orcid.org/0000-0002-3996-7863}{ORCID})
}

}
