% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_normal.R
\name{plot.hmm_mcmc_normal}
\alias{plot.hmm_mcmc_normal}
\title{Plot Method for \code{hmm_mcmc_normal} Objects}
\usage{
\method{plot}{hmm_mcmc_normal}(
  x,
  simulation = FALSE,
  true_means = NULL,
  true_sd = NULL,
  true_mat_T = NULL,
  true_states = NULL,
  show_titles = TRUE,
  ...
)
}
\arguments{
\item{x}{(hmm_mcmc_normal) hmm_mcmc_normal object}

\item{simulation}{(logical) TODO:}

\item{true_means}{(numeric) TODO:}

\item{true_sd}{(numeric) TODO:}

\item{true_mat_T}{(matrix) TODO:}

\item{true_states}{(integer) TODO:}

\item{show_titles}{(logical) if \code{TRUE} then titles are shown for all graphs. By default, \code{TRUE}}

\item{...}{not used}
}
\value{
No return value
}
\description{
Plot Method for \code{hmm_mcmc_normal} Objects
}
\details{
TODO: Here details
}
\examples{
\donttest{
plot(example_hmm_mcmc_normal)
}
}
