% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_poisson.R
\name{plot.hmm_mcmc_poisson}
\alias{plot.hmm_mcmc_poisson}
\title{Plot method for \code{hmm_mcmc_poisson} objects}
\usage{
\method{plot}{hmm_mcmc_poisson}(
  x,
  simulation = FALSE,
  true_betas = NULL,
  true_alpha = NULL,
  true_mat_T = NULL,
  true_states = NULL,
  ...
)
}
\arguments{
\item{x}{(hmm_mcmc_\*) MCMC HMM object}

\item{simulation}{(logical)}

\item{true_betas}{(numeric)}

\item{true_alpha}{(numeric)}

\item{true_mat_T}{(matrix)}

\item{true_states}{(integer)}

\item{...}{not used}
}
\value{
No return value
}
\description{
Plot method for \code{hmm_mcmc_poisson} objects
}
\details{
Here details
}
\examples{
# TODO
}
