% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_poisson.R
\name{posterior_probabilities_poisgamma}
\alias{posterior_probabilities_poisgamma}
\title{Forward-backward algorithm to calculate the posterior probabilities of hidden states in Poisson-Gamma model.}
\usage{
posterior_probabilities_poisgamma(data, pi, mat_T, betas, alpha)
}
\arguments{
\item{data}{(numeric) Poisson data}

\item{pi}{(numeric) prior probability of states}

\item{mat_T}{(matrix) transition probability matrix}

\item{betas}{(numeric) vector with prior rates}

\item{alpha}{(numeric) prior scale}
}
\value{
List with posterior probabilities (TO BE CORRECTED)
}
\description{
Forward-backward algorithm to calculate the posterior probabilities of hidden states in Poisson-Gamma model.
}
\details{
Here details on how the calculation is made
}
\examples{
mat_T <- rbind(c(1-0.01,0.01,0),
               c(0.01,1-0.02,0.01),
               c(0,0.01,1-0.01))
L <- 2^10
betas <- c(0.1, 0.3, 0.5)
alpha <- 1

sim_data <- hmm_simulate_poisgamma_data(L = L,
                                        mat_T = mat_T,
                                        betas = betas,
                                        alpha = alpha)
pi <- sim_data$pi
hmm_poison_data <- sim_data$data
hist(hmm_poison_data, breaks = 100)

# Calculate posterior probabilities of hidden states
post_prob <-  posterior_probabilities_poisgamma(data = hmm_poison_data,
                                                pi = pi,
                                                mat_T = mat_T,
                                                betas = betas,
                                                alpha = alpha)
}
\references{
Here some references
}
