% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_documentation.R
\docType{package}
\name{oHMMed-package}
\alias{oHMMed-package}
\title{oHMMed: HMMs with Ordered Hidden States and Emission Densities}
\description{
Inference using a class of Hidden Markov models (HMMs) called 'oHMMed'(ordered HMM with emission densities \doi{10.1101/2023.06.26.546495}): The 'oHMMed' algorithms identify the number of comparably homogeneous regions within observed sequences with autocorrelation patterns. These are modelled as discrete hidden states; the observed data points are then realisations of continuous probability distributions with state-specific means that enable ordering of these distributions. The observed sequence is labelled according to the hidden states, permitting only neighbouring states that are also neighbours within the ordering of their associated distributions. The parameters that characterise these state-specific distributions are then inferred. Relevant for application to genomic sequences, time series, or any other sequence data with serial autocorrelation.
}
\references{
Inference of Genomic Landscapes using Ordered Hidden Markov Models with Emission Densities (oHMMed)
Claus Vogl, Mariia Karapetiants, Burçin Yildirim, Hrönn Kjartansdottir, Carolin Kosiol, Juraj Bergman, Michal Majka, Lynette Caitlin Mikula,
bioRxiv 2023.06.26.546495; \doi{10.1101/2023.06.26.546495}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LynetteCaitlin/oHMMed}
  \item \url{https://lynettecaitlin.github.io/oHMMed/}
  \item Report bugs at \url{https://github.com/LynetteCaitlin/oHMMed/issues}
}

}
\author{
\strong{Maintainer}: Michal Majka \email{michalmajka@hotmail.com} (\href{https://orcid.org/0000-0002-7524-8014}{ORCID})

Authors:
\itemize{
  \item Lynette Caitlin Mikula \email{lynettecaitlin@gmail.com} (\href{https://orcid.org/0000-0002-0252-4014}{ORCID})
  \item Claus Vogl \email{claus.vogl@vetmeduni.ac.at} (\href{https://orcid.org/0000-0002-3996-7863}{ORCID})
}

}
