% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_documentation.R
\docType{package}
\name{oHMMed-package}
\alias{oHMMed}
\alias{oHMMed-package}
\title{oHMMed: Hidden Markov Models with ordered hidden states and emission densities with applications in population genetics}
\description{
The oHMMed package contains an implementation of Hidden Markov Models with ordered hidden states and emission densities with applications in population genetics.
}
\references{
TODO: Reference to the paper
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LynetteCaitlin/oHMMed}
  \item \url{https://lynettecaitlin.github.io/oHMMed/}
  \item Report bugs at \url{https://github.com/LynetteCaitlin/oHMMed/issues}
}

}
\author{
\strong{Maintainer}: Michal Majka \email{michalmajka@hotmail.com} (\href{https://orcid.org/0000-0002-7524-8014}{ORCID})

Authors:
\itemize{
  \item Lynette Caitlin Mikula (\href{https://orcid.org/0000-0002-0252-4014}{ORCID})
  \item Claus Vogl
}

}
\keyword{internal}
