% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_poisson.R
\name{plot.hmm_mcmc_gamma_poisson}
\alias{plot.hmm_mcmc_gamma_poisson}
\title{Plot Method for \code{hmm_mcmc_gamma_poisson} Objects}
\usage{
\method{plot}{hmm_mcmc_gamma_poisson}(
  x,
  simulation = FALSE,
  true_betas = NULL,
  true_alpha = NULL,
  true_mat_T = NULL,
  true_states = NULL,
  show_titles = TRUE,
  ...
)
}
\arguments{
\item{x}{(hmm_mcmc_gamma_poisson) hmm_mcmc_gamma_poisson object}

\item{simulation}{(logical) TODO:}

\item{true_betas}{(numeric) TODO:}

\item{true_alpha}{(numeric) TODO:}

\item{true_mat_T}{(matrix) TODO:}

\item{true_states}{(integer) TODO:}

\item{show_titles}{(logical) if \code{TRUE} then titles are shown for all graphs. By default, \code{TRUE}}

\item{...}{not used}
}
\value{
No return value
}
\description{
Plot Method for \code{hmm_mcmc_gamma_poisson} Objects
}
\details{
TODO: Here details
}
\examples{
\donttest{
plot(example_hmm_mcmc_gamma_poisson)
}
}
