% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_normal.R
\name{plot.hmm_mcmc_normal}
\alias{plot.hmm_mcmc_normal}
\title{Plot method for \code{hmm_mcmc_normal} objects}
\usage{
\method{plot}{hmm_mcmc_normal}(
  x,
  simulation = FALSE,
  true_means = NULL,
  true_sd = NULL,
  true_mat_T = NULL,
  true_states = NULL,
  ...
)
}
\arguments{
\item{x}{(hmm_mcmc_\*) MCMC HMM object}

\item{simulation}{(logical)}

\item{true_means}{(numeric)}

\item{true_sd}{(numeric)}

\item{true_mat_T}{(matrix)}

\item{true_states}{(integer)}

\item{...}{not used}
}
\value{
No return value
}
\description{
Plot method for \code{hmm_mcmc_normal} objects
}
\details{
Here details
}
\examples{
# TODO

}
